+++ docs/main/changelog.rst
@@ -7,6 +7,10 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+Development version
+-------------------
+- Update providers overriding documentation and rework examples.
+
 3.35.1
 ------
 - Fix minor issues in the providers documentation and examples.